create browser wallet wrapper
